+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 ### Fixed
 
 -   [#509](https://github.com/wordplaydev/wordplay/issues/509) Fixed parsing regression from infinite loop fixes.
+-   [#507](https://github.com/wordplaydev/wordplay/issues/507) Fixed Webpage stream replay bug.
 
 ## 0.10.2 2024-06-29